Add tests for the P2CSelector
