+++ servicetalk-loadbalancer/src/main/java/io/servicetalk/loadbalancer/NewRoundRobinLoadBalancer.java
@@ -217,7 +217,7 @@ final class NewRoundRobinLoadBalancer<ResolvedAddress, C extends LoadBalancedCon
             final Collection<? extends ServiceDiscovererEvent<ResolvedAddress>> events) {
         boolean available = false;
         for (ServiceDiscovererEvent<ResolvedAddress> event : events) {
-            if (host.address.equals(event.address())) {
+            if (host.address().equals(event.address())) {
                 available = true;
                 break;
             }
@@ -276,7 +276,7 @@ final class NewRoundRobinLoadBalancer<ResolvedAddress, C extends LoadBalancedCon
                                 }
                             } else if (UNAVAILABLE.equals(eventStatus)) {
                                 return listWithHostRemoved(oldHostsTyped, host -> {
-                                    boolean match = host.address.equals(addr);
+                                    boolean match = host.address().equals(addr);
                                     if (match) {
                                         host.markClosed();
                                     }
@@ -331,7 +331,7 @@ final class NewRoundRobinLoadBalancer<ResolvedAddress, C extends LoadBalancedCon
         private List<Host<ResolvedAddress, C>> markHostAsExpired(
                 final List<Host<ResolvedAddress, C>> oldHostsTyped, final ResolvedAddress addr) {
             for (Host<ResolvedAddress, C> host : oldHostsTyped) {
-                if (host.address.equals(addr)) {
+                if (host.address().equals(addr)) {
                     // Host removal will be handled by the Host's onClose::afterFinally callback
                     host.markExpired();
                     break;  // because duplicates are not allowed, we can stop iteration
@@ -342,7 +342,7 @@ final class NewRoundRobinLoadBalancer<ResolvedAddress, C extends LoadBalancedCon
 
         private Host<ResolvedAddress, C> createHost(ResolvedAddress addr) {
             // All hosts will share the healthcheck config of the parent RR loadbalancer.
-            Host<ResolvedAddress, C> host = new Host<>(NewRoundRobinLoadBalancer.this.toString(), addr,
+            Host<ResolvedAddress, C> host = new DefaultHost<>(NewRoundRobinLoadBalancer.this.toString(), addr,
                     connectionFactory, linearSearchSpace, healthCheckConfig);
             host.onClose().afterFinally(() ->
                     usedHostsUpdater.updateAndGet(NewRoundRobinLoadBalancer.this, previousHosts -> {
@@ -363,7 +363,7 @@ final class NewRoundRobinLoadBalancer<ResolvedAddress, C extends LoadBalancedCon
 
             // duplicates are not allowed
             for (Host<ResolvedAddress, C> host : oldHostsTyped) {
-                if (host.address.equals(addr)) {
+                if (host.address().equals(addr)) {
                     if (!host.markActiveIfNotClosed()) {
                         // If the host is already in CLOSED state, we should create a new entry.
                         // For duplicate ACTIVE events or for repeated activation due to failed CAS
